server/services/email/requirements.txt
msal==1.24.1
requests==2.31.0
keyring==24.2.0
python-dateutil==2.8.2
flask>=3.1.0
flask-cors>=6.0.0